defer updatin empty UID to insert
