+++ pkg/services/provisioning/datasources/types_test.go
@@ -1,0 +1,15 @@
+package datasources
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestUIDFromNames(t *testing.T) {
+	t.Run("generate safe uid from name", func(t *testing.T) {
+		require.Equal(t, safeUIDFromName("Hello world"), "P64EC88CA00B268E5")
+		require.Equal(t, safeUIDFromName("Hello World"), "PA591A6D40BF42040")
+		require.Equal(t, safeUIDFromName("AAA"), "PCB1AD2119D8FAFB6")
+	})
+}
